requirements.txt
appium-python-client==5.1.2
selenium==4.34.2
opencv-python
pytest
pillow
pyyaml
